requirements.txt
# Object detection and multi-object tracking runtime
ultralytics>=8.3.0,<9.0.0
supervision>=0.23.0,<1.0.0

# Computer vision and image processing
opencv-python>=4.9.0,<5.0.0
Pillow>=10.0.0,<12.0.0

# Scientific and numerical computation
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<2.0.0

# Configuration management
PyYAML>=6.0.1,<7.0.0

# Test runner
pytest>=8.0.0,<9.0.0
